backend/api_app/serializers.py
from rest_framework import serializers
from .models import Api

class ApiSerializer(serializers.ModelSerializer):
    class Meta:
        model = Api
        fields = ['id','first_name','last_name','age','description']
